receiver initialized first
