+++ src/command_pattern/include/command_pattern/ecs_dcm.hpp
@@ -37,6 +37,12 @@ class ECS_DCM
 
         flecs::world dcm_world_;
         BaseReceiver * receiver_;
+        BaseCommand * import_energy_c_;
+        BaseCommand * export_energy_c_;
+        BaseCommand * get_energy_c_;
+        BaseCommand * get_nameplate_c_;
+        BaseCommand * idle_c_;
+
         //BaseInvoker * sim_invoker_;
