+++ src/ecs/include/ecs/der_sim_module.hpp
@@ -4,56 +4,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <flecs.h>
-
+#include "der_components_module.hpp"
 namespace der
 {
-
-class der_components
-{
-    public:
-        struct DERSimulatorEntityTag 
-        {};
-
-        struct SimpleDER
-        {
-            double available_import_energy, available_export_energy, loss_per_second;
-            int loss_rate, import_low, import_high, secs_since_epoch, secs_since_command;
-        };
-
-        enum class CurrentActiveCommand : short //change this enum value to simulate commands
-        {
-            kNoCommand,
-            kImportEnergyCommand,
-            kExportEnergyCommand,
-            kCustomerHasOverriddenCommands
-        };
-
-        enum class Status : short
-        {
-            kImporting,
-            kExporting,
-            kIdle,
-            kCustomerOverride
-        };
-
-        der_components(flecs::world& world_) 
-        {
-            world_.module<der_components>();
-            // Register components 
-            world_.component<Status>();
-            world_.component<SimpleDER>();
-            world_.component<CurrentActiveCommand>();
-            world_.component<DERSimulatorEntityTag>();
-            
-            //world_.set<DER>({1000, 4500, 0, 4500, 0});
-            //create type for simulated der entities
-            auto der_sim_type = world_.type("der_sim_type")
-                .add<Status>()
-                .add<SimpleDER>()
-                .add<CurrentActiveCommand>()
-                .add<DERSimulatorEntityTag>();
-        }
-};
 
 class der_simulator_module 
 {
